List reporter: gather all test results along with elapsed time and file information.

This reporter gathers all results, adding additional information such as test elapsed time, and test filename if available. Very useful for reporting.

See also