Manage test reporting

The job of a reporter is to aggregate the results from files, tests, and expectations and display them in an informative way. Every testtthat function that runs multiple tests provides a reporter argument which you can use to override the default (which is selected by default_reporter()). Typically this will

Reporter

Format

An object of class R6ClassGenerator of length 24.

Details

You only need to use this Reporter object directly if you are creating a new reporter. Currently, creating new Reporters is undocumented, so if you want to create your own, you'll need to make sure that you're familiar with R6 and then need read the source code for a few.

See also

Examples

path <- testthat_example("success")

# The default reporter - doesn't display well in examples because
# it's designed to work in an interactive console.
test_file(path)

#> ✔ |  OK F W S | Context
#> ⠏ |   0       | success✔ |   4   1 1 | success
#> ────────────────────────────────────────────────────────────────────────────────
#> test-success.R:6: skip: you can skip tests if needed
#> This tests hasn't been written yet
#> 
#> test-success.R:10: warning: some tests have warnings
#> NaNs produced
#> ────────────────────────────────────────────────────────────────────────────────
#> 
#> ══ Results ═════════════════════════════════════════════════════════════════════
#> OK:       4
#> Failed:   0
#> Warnings: 1
#> Skipped:  1
#> 
#> You rock!

# Override the default by supplying the name of a reporter
test_file(path, reporter = "minimal")

#> .SW...