SilentReporter {testthat}	R Documentation

Test reporter: gather all errors silently.

Description

This reporter quietly runs all tests, simply gathering all expectations. This is helpful for programmatically inspecting errors after a test run. You can retrieve the results with the expectations() method.

See Also

Other reporters: CheckReporter, DebugReporter, FailReporter, JunitReporter, ListReporter, LocationReporter, MinimalReporter, MultiReporter, ProgressReporter, RStudioReporter, Reporter, StopReporter, SummaryReporter, TapReporter, TeamcityReporter

[Package testthat version 3.2.0 Index]