capture_condition {testthat}	R Documentation

Capture conditions, including messeages, warnings, expectations, and errors.

Description

These functions allow you to capture the side-effects of a function call including printed output, messages and warnings. They are used to evaluate code for expect_output(), expect_message(), expect_warning(), and expect_silent().

Usage

capture_condition(code, entrace = FALSE)

capture_error(code, entrace = FALSE)

capture_expectation(code, entrace = FALSE)

capture_message(code, entrace = FALSE)

capture_warning(code, entrace = FALSE)

capture_messages(code)

capture_warnings(code)

Arguments

code
Code to evaluate

entrace
Whether to add a backtrace to the captured condition.

Value

Singular functions (capture_condition, capture_expectation etc) return a condition object. capture_messages() and capture_warnings return a character vector of message text.

Examples

f <- function() {
  message("First")
  warning("Second")
  message("Third")
}

capture_message(f())
capture_messages(f())

capture_warning(f())
capture_warnings(f())

# Condition will capture anything
capture_condition(f())

[Package testthat version 2.3.1 Index]