expect_cpp_tests_pass {testthat}	R Documentation

Test Compiled Code in a Package

Description

Test compiled code in the package package. See use_catch() for more details.

Usage

expect_cpp_tests_pass(package)

Arguments

package
The name of the package to test.

Note

A call to this function will automatically be generated for you in tests/testthat/test-cpp.R after calling use_catch(); you should not need to manually call this expectation yourself.

[Package testthat version 2.1.1 Index]