expect_equivalent {testthat}	R Documentation

Is an object equal to the expected value, ignoring attributes?

Description

Compares object and expected using all.equal() and check.attributes = FALSE.

Usage

expect_equivalent(
  object,
  expected,
  ...,
  info = NULL,
  label = NULL,
  expected.label = NULL
)

Arguments

object, expected
Computation and value to compare it to.

Both arguments supports limited unquoting to make it easier to generate readable failures within a function or for loop. See quasi_label for more details.

...
Passed on to compare().

info
Extra information to be included in the message. This argument is soft-deprecated and should not be used in new code. Instead see alternatives in quasi_label.

label, expected.label
Used to customise failure messages. For expert use only.

3rd edition

expect_equivalent() is deprecated in the 3rd edition. Instead use expect_equal(ignore_attr = TRUE).

Examples

#' # expect_equivalent() ignores attributes
a <- b <- 1:3
names(b) <- letters[1:3]
## Not run: 
expect_equal(a, b)

## End(Not run)
expect_equivalent(a, b)

[Package testthat version 3.2.0 Index]