Expectation class.

You generate an expectation object with expectation(), and signal an expectation with expect(). If you're making your own expectation function, you'll typically call expect(condition, "failure message"). See the source code for the built-in expectations for details.

expectation(type, message, srcref = NULL)

expect(exp, ..., srcref = NULL)

quasi_label(quo, label = NULL)

is.expectation(x)

Arguments

message
A text description of failure

srcref
Source reference, if known

quo, label
make_label is used to generate informative labels for expressions. quo should be created with rlang::enquo(), making it possible to use unquoting to generate informative labels when running multiple tests in a function or for loop.

x
object to test for class membership

passed
a single logical value indicating whether the test passed (TRUE), failed (FALSE), or threw an error (NA)