local_snapshotter {testthat}	R Documentation

Instantiate local snapshotting context

Description

Needed if you want to run snapshot tests outside of the usual testthat framework For expert use only.

Usage

local_snapshotter(
  snap_dir = NULL,
  cleanup = FALSE,
  fail_on_new = FALSE,
  .env = parent.frame()
)

[Package testthat version 3.1.2 Index]