make_expectation {testthat}	R Documentation

Make an equality test.

Description

This a convenience function to make a expectation that checks that input stays the same.

Usage

make_expectation(x, expectation = "equals")

Arguments

x
a vector of values

expectation
the type of equality you want to test for ("equals", "is_equivalent_to", "is_identical_to")

Examples

x <- 1:10
make_expectation(x)

make_expectation(mtcars$mpg)

df <- data.frame(x = 2)
make_expectation(df)

[Package testthat version 3.2.0 Index]