set_max_fails {testthat}	R Documentation

Set maximum number of test failures allowed before aborting the run

Description

This sets the TESTTHAT_MAX_FAILS env var which will affect both the current R process and any processes launched from it.

Usage

set_max_fails(n)

Arguments

n
Maximum number of failures allowed.

[Package testthat version 3.1.7 Index]