snapshot_accept {testthat}	R Documentation

Snapshot management

Description

snapshot_accept() accepts all modified snapshots.

snapshot_review() opens a Shiny app that shows a visual diff of each modified snapshot. This is particularly useful for whole file snapshots created by expect_snapshot_file().

Usage

snapshot_accept(files = NULL, path = "tests/testthat")

snapshot_review(files = NULL, path = "tests/testthat")

Arguments

files
Optionally, filter effects to snapshots from specified files. This can be a snapshot name (e.g. foo or foo.md), a snapshot file name (e.g. testfile/foo.txt), or a snapshot file directory (e.g. ⁠testfile/⁠).

path
Path to tests.

[Package testthat version 3.2.0 Index]