source_file {testthat}	R Documentation

Source a file, directory, or all helpers.

Description

The expectation is that the files can be sourced in alphabetical order. Helper scripts are R scripts accompanying test scripts but prefixed by helper. These scripts are run once before the tests are run.

Usage

source_file(path, env = test_env(), chdir = TRUE, encoding = "unknown",
  wrap = TRUE)

source_dir(path, pattern = "\\.[rR]$", env = test_env(), chdir = TRUE,
  wrap = TRUE)

source_test_helpers(path = "tests/testthat", env = test_env())

source_test_setup(path = "tests/testthat", env = test_env())

source_test_teardown(path = "tests/testthat", env = test_env())

Arguments

path
Path to tests

env
Environment in which to evaluate code.

chdir
Change working directory to dirname(path)?

encoding
Deprecated

wrap
Automatically wrap all code within test_that()? This ensures that all expectations are reported, even if outside a test block.

pattern
Regular expression used to filter files

[Package testthat version 2.0.1 Index]