test_env {testthat}	R Documentation

Generate default testing environment.

Description

We use a new environment which inherits from globalenv() or a package namespace. In an ideal world, we'd avoid putting the global environment on the search path for tests, but it's not currently possible without losing the ability to load packages in tests.

Usage

test_env(package = NULL)

[Package testthat version 3.1.2 Index]