R package to make testing fun!

Try the example below. Have a look at the references and learn more from function documentation such as expect_that().

Details

Software testing is important, but, in part because it is frustrating and boring, many of us avoid it.

testthat is a new testing framework for R that is easy learn and use, and integrates with your existing workflow.

Options

testthat.use_colours: Should the output be coloured? (Default: TRUE).

testthat.summary.max_reports: The maximum number of detailed test reports printed for the summary reporter (default: 15).

testthat.summary.omit_dots: Omit progress dots in the summary reporter (default: FALSE).

References

Wickham, H (2011). testthat: Get Started with Testing. The R Journal 3/1 5-10. http://journal.r-project.org/archive/2011-1/RJournal_2011-1_Wickham.pdf https://github.com/r-lib/testthat http://adv-r.had.co.nz/Testing.html

Examples

library(testthat)
a <- 9
expect_that(a, is_less_than(10))
expect_lt(a, 10)