ProgressReporter {testthat}	R Documentation

Test reporter: interactive progress bar of errors.

Description

This reporter is a reimagining of SummaryReporter desgined to make the most information available up front, while taking up less space overall. It is the default reporting reporter used by test_dir() and test_file().

Usage

ProgressReporter

Format

An object of class R6ClassGenerator of length 24.

Details

As an additional benefit, this reporter will praise you from time-to-time if all your tests pass.

See Also

Other reporters: CheckReporter, DebugReporter, FailReporter, ListReporter, LocationReporter, MinimalReporter, MultiReporter, Reporter, RstudioReporter, SilentReporter, StopReporter, SummaryReporter, TapReporter, TeamcityReporter

[Package testthat version 2.1.1 Index]