StopReporter {testthat}	R Documentation

Test reporter: stop on error.

Description

The default reporter, executed when expect_that is run interactively. It responds by stop()ping on failures and doing nothing otherwise. This will ensure that a failing test will raise an error.

Usage

StopReporter

Format

An object of class R6ClassGenerator of length 24.

Details

This should be used when doing a quick and dirty test, or during the final automated testing of R CMD check. Otherwise, use a reporter that runs all tests and gives you more context about the problem.

See Also

Other reporters: CheckReporter, DebugReporter, FailReporter, ListReporter, LocationReporter, MinimalReporter, MultiReporter, ProgressReporter, Reporter, RstudioReporter, SilentReporter, SummaryReporter, TapReporter, TeamcityReporter

[Package testthat version 2.1.1 Index]