auto_test_package {testthat}	R Documentation

Watches a package for changes, rerunning tests as appropriate.

Description

Watches a package for changes, rerunning tests as appropriate.

Usage

auto_test_package(pkg = ".", reporter = default_reporter(), hash = TRUE)

Arguments

pkg
path to package

reporter
test reporter to use

hash
Passed on to watch(). When FALSE, uses less accurate modification time stamps, but those are faster for large files.

See Also

auto_test() for details on how method works

[Package testthat version 3.1.2 Index]