expect_is {testthat}	R Documentation

Does an object inherit from a given class?

Description

expect_is() is an older form that uses inherits() without checking whether x is S3, S4, or neither. Instead, I'd recommend using expect_type(), expect_s3_class() or expect_s4_class() to more clearly convey your intent.

Usage

expect_is(object, class, info = NULL, label = NULL)

Arguments

object
Object to test.

Supports limited unquoting to make it easier to generate readable failures within a function or for loop. See quasi_label for more details.

class
Either a character vector of class names, or for expect_s3_class() and expect_s4_class(), an NA to assert that object isn't an S3 or S4 object.

3rd edition

expect_is() is formally deprecated in the 3rd edition.

[Package testthat version 3.1.2 Index]