is_testing {testthat}	R Documentation

Determine testing status

Description

is_testing() determine if code is being run as part of a test

is_parallel() if the test is being run in parallel.

testing_package() gives name of the package being tested.

These are thin wrappers that retrieve the values of environment variables. To avoid creating a run-time dependency on testthat, you can inline the source of these functions directly into your package.

Usage

is_testing()

is_parallel()

testing_package()

[Package testthat version 3.1.2 Index]