not {testthat}	R Documentation

Negate an expectation

Description

This negates an expectation, making it possible to express that you want the opposite of a standard expectation. This function is deprecated and will be removed in a future version.

Usage

not(f)

Arguments

f
an existing expectation function

[Package testthat version 3.1.2 Index]