expect_cpp_tests_pass {testthat}	R Documentation

Do C++ tests past?

Description

Test compiled code in the package package. A call to this function will automatically be generated for you in tests/testthat/test-cpp.R after calling use_catch(); you should not need to manually call this expectation yourself.

Usage

expect_cpp_tests_pass(package)

run_cpp_tests(package)

Arguments

package
The name of the package to test.

[Package testthat version 3.1.2 Index]