teardown_env {testthat}	R Documentation

Run code after all test files

Description

This environment has no purpose other than as a handle for withr::defer(): use it when you want to run code after all tests have been run. Typically, you'll use withr::defer(cleanup(), teardown_env()) immediately after you've made a mess in a setup-*.R file.

Usage

teardown_env()

[Package testthat version 3.1.2 Index]