testthat-package {testthat}	R Documentation

An R package to make testing fun!

Description

Try the example below. Have a look at the references and learn more from function documentation such as test_that().

Options

testthat.use_colours: Should the output be coloured? (Default: TRUE).

testthat.summary.max_reports: The maximum number of detailed test reports printed for the summary reporter (default: 10).

testthat.summary.omit_dots: Omit progress dots in the summary reporter (default: FALSE).

Author(s)

Maintainer: Hadley Wickham hadley@posit.co

Other contributors:

Posit Software, PBC [copyright holder, funder]

R Core team (Implementation of utils::recover()) [contributor]

See Also

Useful links:

https://testthat.r-lib.org

https://github.com/r-lib/testthat

Report bugs at https://github.com/r-lib/testthat/issues

[Package testthat version 3.2.0 Index]