testthat_results {testthat}	R Documentation

Create a testthat_results object from the test results as stored in the ListReporter results field.

Description

Create a testthat_results object from the test results as stored in the ListReporter results field.

Usage

testthat_results(results)

Arguments

results
a list as stored in ListReporter

Value

its list argument as a testthat_results object

See Also

ListReporter

[Package testthat version 3.1.2 Index]